square/pyproject.toml
[tool.poetry]
name = "square"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
websocket-client = "^1.4.1"
numpy = "^1.23.3"
rel = "^0.4.7"

[tool.poetry.dev-dependencies]
black = "^22.10.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
